+++ src/components/Molecules/Tag.jsx
@@ -1,24 +1,24 @@
 import styled from "styled-components";
-import H6 from "../Sangho/H6";
+import H6 from "../Atoms/H6";
 
 const ATag = styled.a`
-  color: black;
-  background-color: #7e7e7e;
-  text-decoration: none;
-  width: 40px;
-  height: 20px;
-  border-radius: 50%;
-  &:hover {
-    background-color: #7e7e7e;
-  }
+	color: black;
+	background-color: #7e7e7e;
+	text-decoration: none;
+	width: 40px;
+	height: 20px;
+	border-radius: 50%;
+	&:hover {
+		background-color: #7e7e7e;
+	}
 `;
 
 export default function Tag({ text }) {
-  return (
-    <>
-      <ATag href={"search/tag?keyword=" + text}>
-        <H6 children={"# " + text}></H6>
-      </ATag>
-    </>
-  );
+	return (
+		<>
+			<ATag href={"search/tag?keyword=" + text}>
+				<H6 children={"# " + text}></H6>
+			</ATag>
+		</>
+	);
 }
