+++ src/index.jsx
@@ -1,75 +1,15 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { createGlobalStyle, ThemeProvider } from "styled-components";
 import App from "./App";
-import theme from "./styles/theme";
 import ResetStyle from "./styles/reset";
+import GlobalVariableStyle from "./styles/global";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-const GlobalStyle = createGlobalStyle`
-/* http://meyerweb.com/eric/tools/css/reset/ 
-   v2.0 | 20110126
-   License: none (public domain)
-*/
-
-/* reset css 적용 */
-html, body, div, span, applet, object, iframe,
-h1, h2, h3, h4, h5, h6, p, blockquote, pre,
-a, abbr, acronym, address, big, cite, code,
-del, dfn, em, img, ins, kbd, q, s, samp,
-small, strike, strong, sub, sup, tt, var,
-b, u, i, center,
-dl, dt, dd, ol, ul, li,
-fieldset, form, label, legend,
-table, caption, tbody, tfoot, thead, tr, th, td,
-article, aside, canvas, details, embed, 
-figure, figcaption, footer, header, hgroup, 
-menu, nav, output, ruby, section, summary,
-time, mark, audio, video {
-	margin: 0;
-	padding: 0;
-	border: 0;
-	font-size: 100%;
-	font: inherit;
-	vertical-align: baseline;
-}
-/* HTML5 display-role reset for older browsers */
-article, aside, details, figcaption, figure, 
-footer, header, hgroup, menu, nav, section {
-	display: block;
-}
-body {
-	line-height: 1;
-}
-ol, ul {
-	list-style: none;
-}
-blockquote, q {
-	quotes: none;
-}
-blockquote:before, blockquote:after,
-q:before, q:after {
-	content: '';
-	content: none;
-}
-table {
-	border-collapse: collapse;
-	border-spacing: 0;
-}
-/* 반응형 디자인을 위해 텍스트를 10px로 수정 */
-font-size: 10px;
-
-:root {
-  --margin-lg: 32px;
-}
-`;
 
 root.render(
 	<React.StrictMode>
 		<ResetStyle />
-
-		<ThemeProvider theme={theme}>
-			<App />
-		</ThemeProvider>
+		<GlobalVariableStyle />
+		<App />
 	</React.StrictMode>
 );
