+++ src/styles/theme.js
@@ -1,3 +1,5 @@
+// @ This file is deprecated.
+
 // 반응형 디자인을 위해 px을 rem으로 전환
 const pixelToRem = (pixel) => `${pixel / 10}rem`;
 
@@ -6,11 +8,7 @@ const fontSize = {
 	h2: pixelToRem(48),
 };
 
-const color = {
-	primary: "green",
-	secondary: "red",
-	ground: "#1e1e1e",
-};
+const color = {};
 
 const padding = {
 	xs: pixelToRem(4),
